requirements.txt
# Core training (CUDA / Threadripper + 3090)
torch>=2.5
transformers>=4.50
peft>=0.15
trl>=0.15
accelerate>=1.5
bitsandbytes>=0.45

# Data
datasets>=4.6
huggingface-hub>=1.5
pillow>=12.0
qwen-vl-utils

# Synthetic data pipeline
flickrapi
imagehash
requests

# Utilities
tqdm
pandas
wandb
pyyaml
